 \name{tspdata} \docType{data} \alias{tspdata} \title{Simulated gene expression data in both matrix and expression set format.} \description{ These data sets can be used to illustrate the TSP functions. The data consist of two simulated data sets "dat", "dat2", a group indicator "grp" for "dat", and two expression sets "eSet1" and "eSet2" where the expression of eSet1 is "dat", the expression of eSet2 is dat2, and the pData for eSet1 consists of "grp".  } \usage{dat} \keyword{datasets}